Erweiterung der Database Views um Beratung, Gewalttat und Taeter

+++ backend/api/database/views.py
@@ -5,6 +5,48 @@ from rest_framework.response import Response
 from .models import *
 from .serializers import *
 
+def type_is_valid(type):
+    return type in ["anfrage", "fall", "beratung", "gewalttat", "taeter"]
+
+def get_serializer(request, type):
+    if type == "anfrage":
+        return AnfrageSerializer(data=request.data)
+    elif type == "fall":
+        return FallSerializer(data=request.data)
+    elif type == "beratung":
+        return BeratungSerializer(data=request.data)
+    elif type == "gewalttat":
+        return GewalttatSerializer(data=request.data)
+    elif type == "taeter":
+        return TaeterSerializer(data=request.data)
+
+def get_data_record(type, pk):
+    if type == "anfrage":
+        try:
+            return Anfrage.objects.get(pk=pk)
+        except Anfrage.DoesNotExist:
+            return None
+    elif type == "fall":
+        try:
+            return Fall.objects.get(pk=pk)
+        except Fall.DoesNotExist:
+            return None
+    elif type == "beratung":
+        try:
+            return Beratung.objects.get(pk=pk)
+        except Beratung.DoesNotExist:
+            return None
+    elif type == "gewalttat":
+        try:
+            return Gewalttat.objects.get(pk=pk)
+        except Gewalttat.DoesNotExist:
+            return None
+    elif type == "taeter":
+        try:
+            return Taeter.objects.get(pk=pk)
+        except Taeter.DoesNotExist:
+            return None
+
 @api_view(["GET"])
 def get_list(request, type):
     """
@@ -13,9 +55,21 @@ def get_list(request, type):
     if type == "anfrage":
         data_record_list = Anfrage.objects.all()
         serializer = AnfrageSerializer(data_record_list, many=True)
-    else:
+    elif type == "fall":
         data_record_list = Fall.objects.all()
         serializer = FallSerializer(data_record_list, many=True)
+    elif type == "beratung":
+        data_record_list = Beratung.objects.all()
+        serializer = BeratungSerializer(data_record_list, many=True)
+    elif type == "gewalttat":
+        data_record_list = Gewalttat.objects.all()
+        serializer = GewalttatSerializer(data_record_list, many=True)
+    elif type == "taeter":
+        data_record_list = Taeter.objects.all()
+        serializer = TaeterSerializer(data_record_list, many=True)
+    else:
+        return Response({"Error": "ungültiges DataRecord"}, status=status.HTTP_400_BAD_REQUEST)
+
     return Response(serializer.data)
 
 @api_view(["POST"])
@@ -23,31 +77,29 @@ def save(request, type):
     """
     Erstellt ein neues DataRecord.
     """
-    if type == "anfrage":
-        serializer = AnfrageSerializer(data=request.data)
-    else:
-        serializer = FallSerializer(data=request.data)
 
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    if not type_is_valid(type):
+        return Response({"Error": "ungültiges DataRecord"}, status=status.HTTP_400_BAD_REQUEST)
+
+    serializer = get_serializer(request, type)
+    if not serializer.is_valid():
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    serializer.save()
+    return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 @api_view(["GET"])
 def get(request, type, pk):
     """
     Gibt ein DataRecord zurück.
     """
-    if type == "anfrage":
-        try:
-            data_record = Anfrage.objects.get(pk=pk)
-        except Anfrage.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-    else:
-        try:
-            data_record = Fall.objects.get(pk=pk)
-        except Fall.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if not type_is_valid(type):
+        return Response({"Error": "ungültiges DataRecord"}, status=status.HTTP_400_BAD_REQUEST)
+    
+    data_record = get_data_record(type, pk)
+    if data_record == None:
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
     serializer = FallSerializer(data_record)
     return Response(serializer.data)
@@ -57,38 +109,33 @@ def update(request, type, pk):
     """
     Überschreibt ein DataRecord.
     """
-    if type == "anfrage":
-        try:
-            data_record = Anfrage.objects.get(pk=pk)
-        except Anfrage.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-    else:
-        try:
-            data_record = Fall.objects.get(pk=pk)
-        except Fall.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if not type_is_valid(type):
+        return Response({"Error": "ungültiges DataRecord"}, status=status.HTTP_400_BAD_REQUEST)
+    
+    data_record = get_data_record(type, pk)
+    if data_record == None:
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
     serializer = FallSerializer(data_record, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    if not serializer.is_valid():
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    serializer.save()
+    return Response(serializer.data)
 
 @api_view(["DELETE"])
 def delete(request, type, pk):
     """
     Löscht ein DataRecord.
     """
-    if type == "anfrage":
-        try:
-            data_record = Anfrage.objects.get(pk=pk)
-        except Anfrage.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-    else:
-        try:
-            data_record = Fall.objects.get(pk=pk)
-        except Fall.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+    
+    if not type_is_valid(type):
+        return Response({"Error": "ungültiges DataRecord"}, status=status.HTTP_400_BAD_REQUEST)
+    
+    data_record = get_data_record(type, pk)
+    if data_record == None:
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
     data_record.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
